requirements.txt
# StateVLA Requirements

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Mamba SSM
mamba-ssm>=1.2.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Data
h5py>=3.7.0
einops>=0.6.0
opencv-python>=4.7.0

# Language Model (CLIP)
ftfy>=6.1.0
regex>=2022.1.18

# Training
tqdm>=4.64.0
wandb>=0.15.0

# Optional: LIBERO benchmark
# libero>=0.1.0

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
